Create GuessingNumber.py
GuessingNumber.py is a number-guessing game. Where users can play this game in two-mode Easy and Hard. In the game, one random number will be generated and the user has to guess. This game will provide the number of attempts as per the level. If anybody wants to improve this game. feel free to make changes.